track_data/requirements.txt
wheel
rauth==0.7.3
sqlalchemy==1.4.20
python-dotenv~=0.18.0
